refactor code to add tests
