+++ src/extension.test.ts
@@ -1,22 +1,24 @@
 import expect from 'expect'
-import { getParamsFromUriPath, matchSentryProject } from './extension'
+import { getParamsFromUriPath, matchSentryProject } from './handler'
 import { Settings } from './settings'
 
 describe('getParamsFromUriPath', () => {
     it('extracts repo and file params', () =>
         expect(
             getParamsFromUriPath('git://github.com/sourcegraph/sourcegraph?264...#web/src/e2e/index.e2e.test.tsx')
-        ).toEqual([{ repo: 'sourcegraph/sourcegraph', file: 'web/src/e2e/index.e2e.test.tsx' }]))
+        ).toEqual({ repo: 'sourcegraph/sourcegraph', file: '#web/src/e2e/index.e2e.test.tsx' }))
 
     it('return empty repo if host is not GitHub', () =>
-        expect(getParamsFromUriPath('git://unknownhost.com/sourcegraph/testrepo#http/req/main.go')).toEqual([
-            { repo: '', file: 'http/req/main.go' },
-        ]))
+        expect(getParamsFromUriPath('git://unknownhost.com/sourcegraph/testrepo#http/req/main.go')).toEqual({
+            repo: '',
+            file: '#http/req/main.go',
+        }))
 
-    it('return file if document has no file format', () =>
-        expect(getParamsFromUriPath('git://github.com/sourcegraph/sourcegraph/testrepo#formatless')).toEqual([
-            { repo: 'sourcegraph/sourcegraph', file: '' },
-        ]))
+    it('return empty file if document has no file format', () =>
+        expect(getParamsFromUriPath('git://github.com/sourcegraph/sourcegraph/testrepo#formatless')).toEqual({
+            repo: 'sourcegraph/sourcegraph',
+            file: '',
+        }))
 })
 
 const projects: Settings['sentry.projects'] = [
@@ -47,14 +49,12 @@ const params = {
 
 describe('matchSentryProject', () => {
     it('extracts repo and file params', () =>
-        expect(matchSentryProject(params, projects)).toEqual([
-            {
-                projectId: '1334031',
-                lineMatches: [
-                    /throw new Error+\(['"]([^'"]+)['"]\)/,
-                    /console\.(warn|debug|info|error|log)\(['"`]([^'"`]+)['"`]\)/,
-                    /log\.(Printf|Print|Println)\(['"]([^'"]+)['"]\)/,
-                ],
-            },
-        ]))
+        expect(matchSentryProject(params, projects)).toEqual({
+            projectId: '1334031',
+            lineMatches: [
+                /throw new Error+\(['"]([^'"]+)['"]\)/,
+                /console\.(warn|debug|info|error|log)\(['"`]([^'"`]+)['"`]\)/,
+                /log\.(Printf|Print|Println)\(['"]([^'"]+)['"]\)/,
+            ],
+        }))
 })
